+++ v2/src/dar_backup/__about__.py
@@ -1,4 +1,6 @@
-__version__ = "1.0.1"
+__version__ = "1.0.2"
+
+__author__ = "Per Jensen"
 
 __license__ = '''Licensed under GNU GENERAL PUBLIC LICENSE v3, see the supplied file "LICENSE" for details.
 THERE IS NO WARRANTY FOR THE PROGRAM, TO THE EXTENT PERMITTED BY APPLICABLE LAW, not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.